requirements.txt
-i https://pypi.org/simple
blinker==1.7.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; platform_system == 'Windows'
dominate==2.9.1; python_version >= '3.4'
flask==3.0.2; python_version >= '3.8'
flask-bootstrap==3.3.7.1
gunicorn==21.2.0; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.3; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
numpy==1.26.4; python_version >= '3.9'
packaging==24.0; python_version >= '3.7'
pandas==2.2.1; python_version >= '3.9'
plotly==5.20.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tenacity==8.2.3; python_version >= '3.7'
tzdata==2024.1; python_version >= '2'
visitor==0.1.3
werkzeug==3.0.2; python_version >= '3.8'
tensorflow
joblib
scikit-learn
psycopg2  
python-dotenv 